renames data to model
related #11
