+++ src/server/model/image.js
